Try to make sure that file ownership works out OK.
On Unix, the best way to do this with a generic container, that I'm
aware of, is to propagate the host ID numbers and use an entrypoint
script to set up all subsequent processes to run as the desired UID/GID.
Fairly gross, but straightforward enough once you've set up the scripts.

On macOS, the above may all be unnecessary, based on the discussion here:

https://stackoverflow.com/questions/43097341/docker-on-macosx-does-not-translate-file-ownership-correctly-in-volumes

It sounds as if our entrypoint may in fact be counterproductive in this
case. This will need some investigation.