Rework args protocol so that we can handle --help
Make it so that the program has to exit with a code of 100 to indicate
that the args analysis succeeded. That way, if a --help option is given
and argparse exits early for us, the frontend driver can easily detect
that fact. More robust than searching for `--help` in the arg text, or
whatever.