backend: use portable compiler flags for the Docker build
The Docker image built on Azure was dying with a SIGILL on my machine,
because Astrometry.Net defaults to building with `-march=native`, which
will aggressively use fancy CPU instructions. This is good for native
builds but not good for Docker where the code may be run on a variety of
CPUs. Fortunately it is pretty straightforward and well-documented how
to adjust these flags.