backend: fix AN "preset" and tweak some parameters
The main change here is tweaking how we choose the Astromery.Net
"preset" value, to bias it to lower values. We also need to make sure
that it's an integer!

Also tweak some other settings:

- Default time limit of 30s: more CPU time here doesn't seem to help,
  which makes sense: in this constrained problem, we'll either find
  a solution pretty quickly, or we just won't
- Bump source threshold to 10 sigma in SEP detection to try to avoid
  getting swamped with low-utility sources
- Change how we calculate the scale_high to be more future-proof

This helps us get about half of the test cases passing, and I think that
from this baseline we can add one or more "try-hard" modes that will get
the rest going.